removed old headpanel, viewing the sign in page now wipes token
