--- src/View/HeadPanel/HeadPanel.css
@@ -1,31 +1,0 @@
-.day {
-    background-color: darkorange !important;
-    color: white !important;
-    width: 10vw;
-    box-shadow: 10;
-}
-
-.button {
-    color: white !important;
-    height: 8vh;
-    padding-top: 10;
-}
-
-.newMeeting {
-    background-color: darkorange !important;
-    color: white !important;
-    box-shadow: 10;
-    width: 13vw;
-}
-
-.headPanelContainer{
-    background-color: #E3E3E3;
-    width: 100vw;
-    height: 10vh;
-}
-
-.headPanelGrid{
-    padding-top: 1vh;
-    padding-left: 1vw;
-    padding-right: 1vw;
-}
